Cap size of depth pyramid

+++ include/RavEngine/DepthPyramid.hpp
@@ -6,11 +6,11 @@
 namespace RavEngine {
 
 	struct DepthPyramid {
-		DepthPyramid(uint32_t width, const std::string_view name = "Depth Pyramid");
+		DepthPyramid(uint16_t width, const std::string_view name = "Depth Pyramid");
 		DepthPyramid() {}
 		RGLTexturePtr pyramidTexture = nullptr;
-		uint32_t dim = 0;	// size of mip 0
-		uint32_t numLevels = 0;
+        uint16_t dim = 0;	// size of mip 0
+        uint8_t numLevels = 0;
 	};
 }
 #endif
